+++ master.py
@@ -7,8 +7,8 @@ Created on Tue Apr  5 10:42:59 2016
 import os
 from multiprocessing import Pool
 
-p=4
-file="main_constant.py"
+p=1
+file="main_variable.py"
 path= "/Users/matthias/Documents/popdyn/botero-model/single_runs/"
 
 def run(line):
@@ -18,33 +18,40 @@ def run(line):
 
 lines=[]
 
-with open(path+"trans_to_plast_new.txt","r") as f:
+with open(path+"base_extinct.txt","r") as f:
     line=f.readline()
     lines.append(line)
     while line:
         line=f.readline()
         lines.append(line)       
-    
-'''Create list of lists of p elements to be used as arguments in pool.map '''
-list1=[]
-list2=[] 
-for i,l in enumerate(lines):
-    list2.append(l[0:-1])
-    if (i % p)==p-1:
-        list1.append(list2)
-        list2=[]
-list1.append(list2)
-
-     
   
-
+if p>1:  
+    '''Create list of lists of p elements to be used as arguments in pool.map '''
+    list1=[]
+    list2=[] 
+    for i,l in enumerate(lines):
+        list2.append(l[0:-1])
+        if (i % p)==p-1:
+            list1.append(list2)
+            list2=[]
+            list1.append(list2) 
+        
+    '''Run p number of processes simultaneously'''
     
-'''Run p number of processes simultaneously'''
-
-a=[]
-for l in list1:
-    if len(l)!=0:
-        pool=Pool(processes=len(l))
-        a.extend(pool.map(run,l))    
-        pool.terminate()
-
+    a=[]
+    for l in list1:
+        if len(l)!=0:
+            try:
+                pool=Pool(processes=len(l))
+                a.extend(pool.map(run,l))    
+                pool.terminate()
+            except:
+                print("Error in: {0}".format(l))
+    #    
+else:
+    for l in lines:
+        if l:
+            try:
+                run(l)
+            except:
+                print("Error in: {0}".format(l))
